+++ models/user/schema.ts
@@ -23,11 +23,12 @@ const schema = new Schema<IUserDoc, IUserModel, IUser>({
     lowercase: true,
   },
   registered_at: {
-    type: Date,
+    type: Number,
     required: true,
+    default: Date.now,
   },
   verified_at: {
-    type: Date,
+    type: Number,
     required: true,
   },
   token_hash: {