+++ plugins/db/helpers/index.ts
@@ -1,3 +1,4 @@
 export { encryptFields, decryptFields } from './encryption';
+export { encodeFields, decodeFields } from './encoder';
 export { default as fieldsPicker } from './fields-picker';
 export { default as hashString } from './hash';